Enum hakari::cli_ops::InitError

#[non_exhaustive]
pub enum InitError {
    ConfigPathExists {
        abs_path: Utf8PathBuf,
    },
    PackageNameExists {
        package_name: String,
        workspace_path: Utf8PathBuf,
    },
    WorkspacePathNotInRoot {
        abs_path: Utf8PathBuf,
        workspace_root: Utf8PathBuf,
    },
    WorkspacePathExists {
        abs_path: Utf8PathBuf,
    },
    Io {
        path: Utf8PathBuf,
        error: Error,
    },
}

An error that occurred while attempting to initialize hakari.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

ConfigPathExists

The configuration path already exists.

Fields

abs_path: Utf8PathBuf

The absolute path of the configuration file.

PackageNameExists

The provided package name already exists.

Fields

package_name: String

The package name that exists.

workspace_path: Utf8PathBuf

The path at which it exists, relative to the root.

WorkspacePathNotInRoot

The provided path is not within the workspace root.

Fields

abs_path: Utf8PathBuf

The absolute workspace path.

workspace_root: Utf8PathBuf

The workspace root.

WorkspacePathExists

The provided workspace directory is non-empty.

Fields

abs_path: Utf8PathBuf

The absolute workspace path.

Io

An IO error occurred while working with the given path.

Fields

path: Utf8PathBuf

The path.

error: Error

The error.