Module hakari::explain

Information about why a dependency is in the workspace-hack.

HakariExplain instances are produced by Hakari::explain. The current API is limited to displaying these instances if the cli-support feature is enabled.

Structs

HakariExplain

The result of a Hakari explain query.

HakariExplainDisplay

A display formatter for HakariExplain.