Module hakari::internals

Access to internal Hakari data structures.

These are provided in case some post-processing needs to be done.

Structs

ComputedInnerValue

The values of ComputedInnerMap.

ComputedValue

The values of a ComputedMap.

OutputKey

A key representing a platform and host/target. Returned by Hakari.

Type Aliases

ComputedInnerMap

A target map or a host map in a ComputedValue.

ComputedMap

The map of all build results computed by Hakari.

OutputMap

The map used by Hakari to generate output TOML.