Type Definition hakari::internals::ComputedInnerMap

type ComputedInnerMap<'g> = BTreeMap<BTreeSet<&'g str>, Vec<(PackageMetadata<'g>, StandardFeatures)>>;

A target map or a host map in a ComputedValue.

The keys are sets of feature names (or empty for no features).

The values are the workspace packages and selected features that cause the key in ComputedMap to be built with the given feature set. They are not defined to be in any particular order.