Type Definition hakari::internals::ComputedMap

type ComputedMap<'g> = BTreeMap<(Option<usize>, &'g PackageId), ComputedValue<'g>>;

The map of all build results computed by Hakari.

The keys are the platform index and the dependency's package ID, and the values are ComputedValue instances that represent the different feature sets this dependency is built with on both the host and target platforms.

The values that are most interesting are the ones where maps have two elements or more: they indicate dependencies with features that need to be unified.

This is an alias for the type of Hakari::computed_map.