Type Alias hakari::internals::OutputMap

pub type OutputMap<'g> = BTreeMap<OutputKey, BTreeMap<&'g PackageId, (PackageMetadata<'g>, BTreeSet<&'g str>)>>;

The map used by Hakari to generate output TOML.

This is a two-level BTreeMap, where:

the top-level keys are OutputKey instances.

the inner map is keyed by dependency PackageId instances, and the values are the corresponding PackageMetadata for this dependency, and the set of features enabled for this package.

This is an alias for the type of Hakari::output_map.

Aliased Type

struct OutputMap<'g> { /* private fields */ }