hakari
Module verify

Code related to ensuring that hakari works properly.

Verification algorithm

By default, Hakari runs in “generate mode”: the goal of this mode is to update an existing Hakari package’s TOML. In this mode, the Hakari package is always omitted from consideration and added to the omitted packages.

In verify mode, the goal is to ensure that Cargo builds actually produce a unique set of features for every third-party dependency. In this mode, instead of being omitted, the Hakari package is always included in feature resolution (with default features), through the features_only argument to CargoSet::new. If, in the result, the output_map is empty, then features were unified.

Structs

VerifyErrors

Context for errors returned by HakariBuilder::verify.

VerifyErrorsDisplay

A display formatter for VerifyErrors.