hakari
Struct Hakari

#[non_exhaustive]
pub struct Hakari<'g> {
    pub output_map: OutputMap<'g>,
    pub computed_map: ComputedMap<'g>,
    /* private fields */
}

The result of a Hakari computation.

This contains all the data required to generate a workspace package.

Produced by HakariBuilder::compute.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

output_map: OutputMap<'g>

The map built by Hakari of dependencies that need to be unified.

This map is used to construct the TOML output. Public access is provided in case some post-processing needs to be done.

computed_map: ComputedMap<'g>

The complete map of dependency build results built by Hakari.

This map is not used to generate the TOML output.

Implementations

impl<'g> Hakari<'g>

pub fn builder(&self) -> &HakariBuilder<'g>

Returns the HakariBuilder used to create this instance.

pub fn read_toml(&self) -> Option<Result<HakariCargoToml, CargoTomlError>>

Reads the existing TOML file for the Hakari package from disk, returning a HakariCargoToml.

This can be used with to_toml_string to manage the contents of the given TOML file on disk.

Returns an error if there was an issue reading the TOML file from disk, or None if the builder’s hakari_package is None.

pub fn write_toml( &self, options: &HakariOutputOptions, out: impl Write, ) -> Result<(), TomlOutError>

Writes [dependencies] and other Cargo.toml lines to the given fmt::Write instance.

&mut String and fmt::Formatter both implement fmt::Write.

pub fn toml_name_map(&self) -> AHashMap<Cow<'g, str>, PackageMetadata<'g>>

Returns a map of dependency names as present in the workspace-hack’s Cargo.toml to their corresponding PackageMetadata.

Packages which have one version are present as their original names, while packages with more than one version have a hash appended to them.

pub fn explain( &self, package_id: &'g PackageId, ) -> Result<HakariExplain<'g, '_>, Error>

Returns a HakariExplain, which can be used to print out why a specific package is in the workspace-hack’s Cargo.toml.

Returns an error if the package ID was not found in the output.

pub fn to_toml_string( &self, options: &HakariOutputOptions, ) -> Result<String, TomlOutError>

A convenience method around write_toml that returns a new string with Cargo.toml lines.

The returned string is guaranteed to be valid TOML, and can be provided to a HakariCargoToml obtained from read_toml.