Struct hakari::HakariBuilder

pub struct HakariBuilder<'g> { /* fields omitted */ }

Configures and constructs Hakari instances.

This struct provides a number of options that determine how Hakari instances are generated.

Implementations

impl<'g> HakariBuilder<'g>

pub fn from_summary(
graph: &'g PackageGraph,
summary: &HakariBuilderSummary
) -> Result<Self, Error>

Constructs a HakariBuilder from a PackageGraph and a serialized summary.

Requires the summaries feature to be enabled.

Returns an error if the summary references a package that’s not present, or if there was some other issue while creating a HakariBuilder from the summary.

impl<'g> HakariBuilder<'g>

pub fn new(
graph: &'g PackageGraph,
hakari_id: Option<&PackageId>
) -> Result<Self, Error>

Creates a new HakariBuilder instance from a PackageGraph.

The Hakari package itself is usually present in the workspace. If so, specify its package ID, otherwise pass in None. If specified, this package is marked as special: see the documentation for set_verify_mode for more.

Returns an error if a Hakari package ID is specified but it isn’t known to the graph, or isn’t in the workspace.

pub fn graph(&self) -> &'g PackageGraph

Returns the PackageGraph used to construct this Hakari instance.

pub fn hakari_package(&self) -> Option<&PackageMetadata<'g>>

Returns the Hakari package, or None if it wasn’t passed into new.

pub fn read_toml(&self) -> Option<Result<HakariCargoToml, CargoTomlError>>

Reads the existing TOML file for the Hakari package from disk, returning a HakariCargoToml.

This can be used with Hakari::to_toml_string to manage the contents of the Hakari package’s TOML file on disk.

Returns an error if there was an issue reading the TOML file from disk, or None if this builder was created without a Hakari package.

pub fn set_platforms(
&mut self,
platforms: impl IntoIterator<Item = Platform>
) -> &mut Self

Sets a list of platforms for hakari to use.

By default, hakari unifies features across all platforms. This may not always be desired, so it is possible to set a list of platforms. If the features for a particular dependency only need to be unified on some platforms, hakari will output platform-specific instructions.

Call set_platforms with an empty list to reset to default behavior.

pub fn platforms(&self) -> &[Platform]

Returns the platforms set through set_platforms, or an empty list if no platforms are set.

pub fn set_resolver_version(
&mut self,
version: CargoResolverVersion
) -> &mut Self

Sets the Cargo resolver version.

By default, HakariBuilder uses version 1 of the Cargo resolver. For more about Cargo resolvers, see the documentation for CargoResolverVersion.

pub fn resolver_version(&self) -> CargoResolverVersion

Returns the current Cargo resolver version.

pub fn add_omitted_packages<'b>(
&mut self,
omitted_packages: impl IntoIterator<Item = &'b PackageId>
) -> Result<&mut Self, Error>

Adds packages to not consider while performing unification.

Users may wish to not consider certain packages while figuring out the unified feature set. Setting this option prevents those packages from being considered.

Practically, this means that:

If a workspace package is specified, Cargo build simulations for it will not be run.

If a third-party package is specified, it will not be present in the output, nor will any features enabled by it that aren’t enabled any other way.

Returns an error if any package IDs specified aren’t known to the graph.

pub fn omitted_packages<'b>(
&'b self
) -> impl Iterator<Item = &'g PackageId> + 'b

Returns the currently omitted packages.

If verify_mode is currently false (the default), also returns the Hakari package if specified. This is because the Hakari package is treated as omitted by the algorithm.

pub fn omits_package(&self, package_id: &PackageId) -> Result<bool, Error>

Returns true if a package ID is currently omitted from the set.

If verify_mode is currently false (the default), also returns true for the Hakari package if specified. This is because the Hakari package is treated as omitted by the algorithm.

Returns an error if this package ID isn’t known to the underlying graph.

pub fn set_verify_mode(&mut self, verify_mode: bool) -> &mut Self

If set to true, runs Hakari in verify mode.

By default, Hakari runs in generate mode: the goal of this mode is to update an existing Hakari package’s TOML. In this mode, the Hakari package is always omitted from consideration and added to the omitted packages.

In verify mode, the goal is to ensure that Cargo builds actually produce a unique set of features. In this mode, instead of being omitted, the Hakari package is always included in feature resolution (default features), through the features_only argument to CargoSet::new. If, in the result, output_map is empty, then features were unified.

Setting this to true has no effect if the Hakari package is not specified at construction time.

pub fn verify_mode(&self) -> bool

Returns the current value of verify_mode.

pub fn set_unify_target_host(
&mut self,
unify_target_host: UnifyTargetHost
) -> &mut Self

Whether to unify feature sets across target and host platforms.

By default, hakari does not perform any unification across the target and host platforms. This means that if a dependency is a target (regular) dependency with one set of features, and a host (build) dependency with a different set of features, the two are treated separately.

For more information about this option, see the documentation for UnifyTargetHost.

pub fn unify_target_host(&self) -> UnifyTargetHost

Returns the current value of unify_target_host.

pub fn set_unify_all(&mut self, unify_all: bool) -> &mut Self

Whether to unify feature sets for all dependencies.

By default, Hakari only produces output for dependencies that are built with more than one feature set. If set to true, Hakari will produce outputs for all dependencies, including those that don’t need to be unified.

This is rarely needed in production, and is most useful for testing and debugging scenarios.

pub fn unify_all(&self) -> bool

Returns the current value of unify_all.

pub fn compute(self) -> Hakari<'g>

Computes the Hakari for this builder.

impl<'g> HakariBuilder<'g>

Helpers for property testing

The methods in this section allow random instances of a HakariBuilder to be generated, for use in property-based testing scenarios.

Requires the proptest1 feature to be enabled.

pub fn prop010_strategy(
graph: &'g PackageGraph,
hakari_id_strategy: impl Strategy<Value = Option<&'g PackageId>> + 'g
) -> impl Strategy<Value = HakariBuilder<'g>> + 'g

Returns a Strategy that generates random HakariBuilder instances based on this graph.

Requires the proptest1 feature to be enabled.

Panics

Panics if:

there are no packages in this PackageGraph, or

hakari_id is specified but it isn’t known to the graph, or isn’t in the workspace.

impl<'g> HakariBuilder<'g>

pub fn to_summary(&self) -> Result<HakariBuilderSummary, TargetSpecError>

Converts this HakariBuilder to a serializable summary.

Requires the summaries feature to be enabled.

Returns an error if there are any custom platforms. Serializing custom platforms is currently unsupported.