Module hakari::summaries

Manage configuration and generate summaries for hakari.

Requires the cli-support feature to be enabled.

Structs

HakariBuilderSummary

A HakariBuilder in serializable form. This forms the configuration file format for hakari.

HakariConfig

Configuration for hakari.

OutputOptionsSummary

Options for hakari TOML output, in serializable form.

Statics

DEFAULT_CONFIG_PATH

The location of the configuration used by cargo hakari, relative to the workspace root.

FALLBACK_CONFIG_PATH

The fallback location, used by previous versions of cargo hakari.